+++ extension/src/text.ts
@@ -4,3 +4,11 @@ export function capitalizeString(predicate: string): string {
   }
   return "";
 }
+
+export function* searchSubstring(haystack: string, needle: string) {
+  let index: number = haystack.indexOf(needle);
+  while (index >= 0) {
+    yield index;
+    index = haystack.indexOf(needle, index + needle.length);
+  }
+}
